Extract shared E2E helpers into testutil package
Move generic Kubernetes test helpers from virtualmcp/helpers.go into
a shared test/e2e/thv-operator/testutil/ package so both virtualmcp
and acceptance_tests can reuse them without duplication.

Extracted helpers:
- CheckPodsReady: checks pod readiness by labels
- GetPodLogs: retrieves container logs via clientset
- WaitForMCPServerRunning: polls MCPServer phase
- CreateNodePortService / GetNodePort: NodePort access

The virtualmcp package delegates its private checkPodsReady and
getPodLogs to testutil, keeping its public API stable.
