+++ test/e2e/thv-operator/acceptance_tests/helpers.go
@@ -17,12 +17,10 @@ import (
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/util/intstr"
 	"k8s.io/utils/ptr"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
-	mcpv1alpha1 "github.com/stacklok/toolhive/cmd/thv-operator/api/v1alpha1"
 	"github.com/stacklok/toolhive/test/e2e/images"
 )
 
@@ -98,85 +96,6 @@ func CleanupRedis(ctx context.Context, c client.Client, namespace string) {
 	_ = c.Delete(ctx, &corev1.Service{
 		ObjectMeta: metav1.ObjectMeta{Name: "redis", Namespace: namespace},
 	})
-}
-
-// WaitForMCPServerRunning waits for an MCPServer to reach the Running phase.
-func WaitForMCPServerRunning(
-	ctx context.Context,
-	c client.Client,
-	name, namespace string,
-	timeout, pollingInterval time.Duration,
-) {
-	gomega.Eventually(func() error {
-		server := &mcpv1alpha1.MCPServer{}
-		if err := c.Get(ctx, types.NamespacedName{
-			Name:      name,
-			Namespace: namespace,
-		}, server); err != nil {
-			return err
-		}
-		if server.Status.Phase == mcpv1alpha1.MCPServerPhaseFailed {
-			return gomega.StopTrying(fmt.Sprintf("MCPServer %s failed: %s", name, server.Status.Message))
-		}
-		if server.Status.Phase != mcpv1alpha1.MCPServerPhaseRunning {
-			return fmt.Errorf("MCPServer %s not ready, phase: %s", name, server.Status.Phase)
-		}
-		return nil
-	}, timeout, pollingInterval).Should(gomega.Succeed())
-}
-
-// CreateNodePortService creates a NodePort service targeting the MCPServer proxy pods.
-func CreateNodePortService(
-	ctx context.Context,
-	c client.Client,
-	serverName, namespace string,
-) {
-	service := &corev1.Service{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      serverName + "-nodeport",
-			Namespace: namespace,
-		},
-		Spec: corev1.ServiceSpec{
-			Type: corev1.ServiceTypeNodePort,
-			Selector: map[string]string{
-				"app.kubernetes.io/name":     "mcpserver",
-				"app.kubernetes.io/instance": serverName,
-			},
-			Ports: []corev1.ServicePort{
-				{Port: 8080, TargetPort: intstr.FromInt32(8080), Protocol: corev1.ProtocolTCP},
-			},
-		},
-	}
-	gomega.Expect(c.Create(ctx, service)).To(gomega.Succeed())
-}
-
-// GetNodePort waits for a NodePort service to get a port assigned and returns it.
-func GetNodePort(
-	ctx context.Context,
-	c client.Client,
-	serviceName, namespace string,
-	timeout, pollingInterval time.Duration,
-) int32 {
-	var nodePort int32
-
-	gomega.Eventually(func() error {
-		service := &corev1.Service{}
-		if err := c.Get(ctx, types.NamespacedName{
-			Name:      serviceName,
-			Namespace: namespace,
-		}, service); err != nil {
-			return err
-		}
-		for _, port := range service.Spec.Ports {
-			if port.NodePort > 0 {
-				nodePort = port.NodePort
-				return nil
-			}
-		}
-		return fmt.Errorf("no NodePort assigned yet on service %s", serviceName)
-	}, timeout, pollingInterval).Should(gomega.Succeed())
-
-	return nodePort
 }
 
 // SendToolCall sends a JSON-RPC tools/call request and returns the HTTP status code and body.
